requirements.txt
crewai
groq
yfinance
requests
beautifulsoup4
python-dotenv
apscheduler
pandas
numpy
httpx
lxml
openai
pydantic
typing-extensions
urllib3
python-dateutil
pytz
tzdata
playwright